+++ packages/javascript/bh-shared-ui/src/views/ZoneManagement/Details/DetailsList.tsx
@@ -27,7 +27,8 @@ import { UseQueryResult } from 'react-query';
 import { SortableHeader } from '../../../components';
 import { SortOrder } from '../../../types';
 import { cn } from '../../../utils';
-import { SelectedHighlight, getListHeight, isSelector, isTag, itemSkeletons } from './utils';
+import { SelectedHighlight, getListHeight, isSelector, isTag } from './utils';
+import { itemSkeletons } from '../utils'
 
 const getCountElement = (listItem: AssetGroupTag | AssetGroupTagSelector): React.ReactNode => {
     if (listItem.counts === undefined) {
