+++ packages/javascript/bh-shared-ui/src/views/ZoneManagement/ZoneManagement.tsx
@@ -36,7 +36,6 @@ import { OWNED_ID, TIER_ZERO_ID } from './utils';
 
 const Details = React.lazy(() => import('./Details/Details'));
 const Save = React.lazy(() => import('./Save'));
-const Summary = React.lazy(() => import('./Summary/Summary'));
 
 const detailsPaths = [
     ROUTE_ZONE_MANAGEMENT_TIER_DETAILS,
